x/airsettle/types/message_submit_validator_vote_test.go
package types

import (
	"testing"

	"airsettle/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSubmitValidatorVote_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSubmitValidatorVote
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSubmitValidatorVote{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSubmitValidatorVote{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
